requirements.txt
requests>=2.31.0
schedule>=1.2.0
pytz>=2023.3
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
feedparser>=6.0.11
python-dateutil==2.8.2 